# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is swagger",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/createRoom": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Agent uses this route to create chat room",
                "operationId": "create-chat-room",
                "parameters": [
                    {
                        "type": "string",
                        "format": "userID",
                        "description": "id of user",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "fio",
                        "description": "users name",
                        "name": "fio",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getFaqs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faq"
                ],
                "summary": "get FAQ",
                "operationId": "get-faq",
                "parameters": [
                    {
                        "description": "search question to faq",
                        "name": "filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FaqFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faq"
                ],
                "summary": "Update FAQ",
                "operationId": "update-faq",
                "parameters": [
                    {
                        "description": "Change the FAQ from db",
                        "name": "FAQ_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Faq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faq"
                ],
                "summary": "Create FAQ",
                "operationId": "create-faq",
                "parameters": [
                    {
                        "description": "Create FAQ and save it in db",
                        "name": "FAQ_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Faq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faq"
                ],
                "summary": "Delete FAQ",
                "operationId": "delete-faq",
                "parameters": [
                    {
                        "type": "string",
                        "format": "id",
                        "description": "id of which faq we must delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/joinRoom": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Supports choose chat to join",
                "operationId": "join-chat-room",
                "parameters": [
                    {
                        "type": "string",
                        "format": "fio",
                        "description": "supports name",
                        "name": "fio",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "roomID",
                        "description": "id of room which support must join or if want get max users and there id set it 0)",
                        "name": "room_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Faq": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "key": {
                    "type": "string",
                    "example": "1"
                },
                "question": {
                    "type": "string",
                    "example": "1"
                },
                "response": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "models.FaqFilters": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "key": {
                    "type": "string",
                    "example": "1"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "pageLimit": {
                    "type": "integer",
                    "example": 1
                },
                "product": {
                    "type": "string",
                    "example": "1"
                },
                "project": {
                    "type": "string",
                    "example": "1"
                },
                "question": {
                    "type": "string",
                    "example": "1"
                },
                "search": {
                    "type": "string",
                    "example": "1"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}